int --> size_t

+++ Libft/ft_strlen.c
@@ -1,6 +1,6 @@
-int	ft_strlen(char *str)
+size_t	ft_strlen(char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i] != '\0')
